feature(recipe): Check when user validates a new recipe that it is not already existing in the database
As it is really hard to know exactly the equality between two recipes (mainly because of typos and OCR), choose to focus on ingredients. Except from some basic ingredients like salt or olive oil, they must be around 20% close to be consider as potentially similar.

Add fuse to detect if the title is close too.

Only if those criteria are filled, the recipe will be considered as similar to another one.