feat(Recipe): add a dialog in case user try to use a tag that doesn't exist while adding a recipe
Even if dropdown guide user when he enter a tag to use on its new recipe, he can typo or decide to use a brand new tag.
Detect if it is a typo with Fuse and return the detected tag that has the highest score (i.e. the closest). Propose to the user to use this closes tag or to add the one he wrote.

No more use ItemDialog.tsx for tag (even in TagsSettings.tsx) because it add too much complexity. To see when doing the same behavior on ingredients if it is still possible to mutualize this.