feat(Recipe): add dialogs related to user typing an ingredient name unknownw
Do like tags: Fist try to search in database if it seems like a typo. If it is, return a dialog to the user asking him if he want to use the existing ingredient we found of if he wants to add anyway.

Refactor dialogs so that ItemDialog stay generic but add the dialog for the ingredient similarity