feat(Recipe): add rendering of nutrition datas
In order to render it properly, create `RecipeNutrition` which will render the nutrition part.

It is composed of `NutritionEmptyState` and `NutritionTable`.
One that render when the nutrition data is undefined and one when it is not.

A `NutritionTable` is composed of multiples `NutritionRow`.

For each of the molecules components, make sure to handle modes:
* Read-Only
* Edit
* Add manually

OCR will be done later